move add-hook kill-buffer-hook after defun
kill-buffer-hook as an extremely dangerous hook to add to in the even
that something fails. The fact that symbol that was added was void at
the time of binding was one of the underlying causes of this emacs bug
https://debbugs.gnu.org/cgi/bugreport.cgi?bug=43190.

Therefore I moved the call to add-hook after the definition of
jsons-remove-buffer so that it is impossible for the symbol to be
unbound in the event that something goes wrong in between the addition
of the symbol to the hook and the binding of the symbol to a function.
Instead, if a failure occures before adding the hook, the hook will
never be added, avoiding a nasty state where kill-buffer fails, taking
down quite a bit of emacs functionality with it.